# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Error {
  error: string;
  details?: string;
}

export type RiskLevel = typeof RiskLevel[keyof typeof RiskLevel];


export const RiskLevel = {
  CRITICAL: 'CRITICAL',
  HIGH: 'HIGH',
  MEDIUM: 'MEDIUM',
  STABLE: 'STABLE',
} as const;

export type PatientStatus = typeof PatientStatus[keyof typeof PatientStatus];


export const PatientStatus = {
  QUEUED: 'QUEUED',
  REVIEWED: 'REVIEWED',
} as const;

export interface Patient {
  id: string;
  full_name: string;
  age: number;
  gender: string;
  chief_complaint: string;
  risk_level: RiskLevel;
  status: PatientStatus;
  created_at: string;
}

export interface CreatePatientRequest {
  /** @minLength 1 */
  full_name: string;
  /**
     * @minimum 0
     * @maximum 130
     */
  age: number;
  /** @minLength 1 */
  gender: string;
  chief_complaint: string;
  risk_level: RiskLevel;
  status: PatientStatus;
}

export interface Biomarker {
  name: string;
  value: string;
  normal_range: string;
  status: string;
}

export interface SoapNote {
  subjective: string;
  objective: string;
  assessment: string;
  plan: string;
}

export interface LabReport {
  id: string;
  patient_id: string;
  filename: string;
  patient_name: string;
  risk_level: RiskLevel;
  overall_summary: string;
  biomarkers: Biomarker[];
  soap_note: SoapNote;
  patient_explanation: string;
  analysis_mode: string;
  created_at: string;
}

export interface GenerateReferralRequest {
  patient_id: string;
}

export interface Referral {
  id: string;
  patient_id: string;
  referral_text: string;
  analysis_mode: string;
  created_at: string;
}

/**
 * Invalid request
 */
export type BadRequestResponse = Error;

/**
 * Resource not found
 */
export type NotFoundResponse = Error;

/**
 * AI service failure or malformed response
 */
export type AIServiceErrorResponse = Error;

export type AnalyzeLabReportBody = {
  file: Blob | File;
  patient_id?: string;
};