# lib/api-zod/src/generated/types/biomarker.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Biomarker {
  name: string;
  value: string;
  normal_range: string;
  status: string;
}