# lib/api-zod/src/generated/types/createPatientRequest.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PatientStatus } from './patientStatus';
import type { RiskLevel } from './riskLevel';

export interface CreatePatientRequest {
  /** @minLength 1 */
  full_name: string;
  /**
     * @minimum 0
     * @maximum 130
     */
  age: number;
  /** @minLength 1 */
  gender: string;
  chief_complaint: string;
  risk_level: RiskLevel;
  status: PatientStatus;
}