# lib/api-zod/src/generated/types/labReport.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Biomarker } from './biomarker';
import type { RiskLevel } from './riskLevel';
import type { SoapNote } from './soapNote';

export interface LabReport {
  id: string;
  patient_id: string;
  filename: string;
  patient_name: string;
  risk_level: RiskLevel;
  overall_summary: string;
  biomarkers: Biomarker[];
  soap_note: SoapNote;
  patient_explanation: string;
  analysis_mode: string;
  created_at: Date;
}