# lib/api-zod/src/generated/types/patient.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PatientStatus } from './patientStatus';
import type { RiskLevel } from './riskLevel';

export interface Patient {
  id: string;
  full_name: string;
  age: number;
  gender: string;
  chief_complaint: string;
  risk_level: RiskLevel;
  status: PatientStatus;
  created_at: Date;
}