# lib/api-zod/src/generated/types/patientStatus.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export type PatientStatus = typeof PatientStatus[keyof typeof PatientStatus];


export const PatientStatus = {
  QUEUED: 'QUEUED',
  REVIEWED: 'REVIEWED',
} as const;