# lib/api-zod/src/generated/types/referral.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Referral {
  id: string;
  patient_id: string;
  referral_text: string;
  analysis_mode: string;
  created_at: Date;
}