# lib/api-zod/src/generated/types/riskLevel.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export type RiskLevel = typeof RiskLevel[keyof typeof RiskLevel];


export const RiskLevel = {
  CRITICAL: 'CRITICAL',
  HIGH: 'HIGH',
  MEDIUM: 'MEDIUM',
  STABLE: 'STABLE',
} as const;