# lib/api-zod/src/generated/types/soapNote.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SoapNote {
  subjective: string;
  objective: string;
  assessment: string;
  plan: string;
}